Move Cake\Core\ServiceProvider to the container package
ServiceProvider only ever depended on Cake\Container classes
(AbstractServiceProvider, BootableServiceProviderInterface,
ContainerInterface) - it's a thin CakePHP-flavored convenience layer
over the container package's own service provider base class, not
something that belongs to Core. Moved to Cake\Container\ServiceProvider
alongside it, dropping the now-redundant ContainerInterface import
since it's the same namespace.

Test file and the two TestApp service provider fixtures moved/updated
to match. No composer.json changes needed - cakephp/core already
requires cakephp/container for its other Container-typed interfaces
(ContainerApplicationInterface, PluginInterface, BasePlugin), and
cakephp/container's PSR-4 autoload already covers the new file.

Full test suite passes (10,183 tests, same pre-existing
stdin-availability failure); phpcs clean, phpstan identical to the
pre-move baseline.
