+++ tests/TestCase/Container/ServiceProviderTest.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
  * @since         4.2.0
  * @license       https://opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Test\TestCase\Core;
+namespace Cake\Test\TestCase\Container;
 
 use Cake\Container\Container;
 use Cake\TestSuite\TestCase;
